Moved ACRCloud to plugin. Remove ACRCloud deps from requirements Update README Update CHANGELOG Update Version Number
