+++ mporg/__init__.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-VERSION = "0.2a1"
+VERSION = "0.2a2"
 
 CONFIG_DIR = Path.home() / ".MP3ORG"
 LOG_DIR = CONFIG_DIR / "Logs"
