+++ requirements.txt
@@ -1,16 +1,16 @@
 diskcache~=5.6.3
-ftfy~=6.1.1
-musicbrainzngs~=0.7.1
+#ftfy~=6.1.1
+#musicbrainzngs~=0.7.1
 mutagen~=1.47.0
-pyacoustid~=1.3.0
+#pyacoustid~=1.3.0
 requests~=2.31.0
 urllib3~=2.0.4
 LyricsSearcher @ git+https://github.com/Drag-3/LyricsSearcher.git
 tqdm~=4.66.1
 rich~=13.5.3
 # Linux requirements
-pyacrcloud~=1.0.1; sys_platform == 'linux'
+#pyacrcloud~=1.0.1; sys_platform == 'linux'
 
 # Windows requirements
-pyacrcloud @ https://github.com/acrcloud/acrcloud_sdk_python.git ; sys_platform == 'win32'
+#pyacrcloud @ https://github.com/acrcloud/acrcloud_sdk_python.git ; sys_platform == 'win32'
